atc: structure: wire artifacts in get/put/check/task step
...rather than in the worker.Client. there was some awkwardness in the
ContainerSpec. task/put inputs and (at times) images would be passed in
as runtime.Artifacts for the sole purpose of converting them to Sources
(which are also stored on the ContainerSpec). there was a TODO in the
worker.Client to this effect:

// don't modify spec inside here, Specs don't change after you write them

I considered just adding this responsibility to the worker.Pool (since
it already knows where to find volumes), but that feels like giving too
much responsibility to the Pool (and would require adding back the dummy
`compression` to the API's worker.Pool)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>